udh ganti tipe data di database

+++ Modules/PurchasesReturn/Database/Migrations/2021_08_08_222612_create_purchase_return_details_table.php
@@ -20,12 +20,12 @@ class CreatePurchaseReturnDetailsTable extends Migration
             $table->string('product_name');
             $table->string('product_code');
             $table->integer('quantity');
-            $table->integer('price');
-            $table->integer('unit_price');
-            $table->integer('sub_total');
-            $table->integer('product_discount_amount');
+            $table->unsignedBigInteger('price');
+            $table->unsignedBigInteger('unit_price');
+            $table->unsignedBigInteger('sub_total');
+            $table->unsignedBigInteger('product_discount_amount');
             $table->string('product_discount_type')->default('fixed');
-            $table->integer('product_tax_amount');
+            $table->unsignedBigInteger('product_tax_amount');
             $table->foreign('purchase_return_id')->references('id')
                 ->on('purchase_returns')->cascadeOnDelete();
             $table->foreign('product_id')->references('id')
